requirements.txt
functions-framework==3.*
google-auth==2.*
google-api-python-client==2.*
requests==2.*
